+++ src/EnergizeB.js
@@ -20,8 +20,7 @@ const Me             = imports.misc.extensionUtils.getCurrentExtension();
 const utils          = Me.imports.src.utils;
 
 //////////////////////////////////////////////////////////////////////////////////////////
-// This effect lets your windows be carried to the realm of dreams by some little       //
-// fairies. It's implemented with several overlaid grids of randomly moving points.     //
+// This effect looks a bit like the transporter effect from TNG.                        //
 //////////////////////////////////////////////////////////////////////////////////////////
 
 // The shader class for this effect is registered further down in this file.
@@ -30,7 +29,7 @@ let Shader = null;
 // The effect class is completely static. It can be used to get some metadata (like the
 // effect's name or supported GNOME Shell versions), to initialize the respective page of
 // the settings dialog, as well as to create the actual shader for the effect.
-var Wisps = class Wisps {
+var EnergizeB = class EnergizeB {
 
   // ---------------------------------------------------------------------------- metadata
 
@@ -43,13 +42,13 @@ var Wisps = class Wisps {
   // required. It should match the prefix of the settings keys which store whether the
   // effect is enabled currently (e.g. the '*-close-effect').
   static getNick() {
-    return 'wisps';
+    return 'energize-b';
   }
 
   // This will be shown in the sidebar of the preferences dialog as well as in the
   // drop-down menus where the user can choose the effect.
   static getLabel() {
-    return 'Wisps';
+    return 'Energize B';
   }
 
   // -------------------------------------------------------------------- API for prefs.js
@@ -60,17 +59,18 @@ var Wisps = class Wisps {
   static initPreferences(dialog) {
 
     // Add the settings page to the builder.
-    dialog.getBuilder().add_from_resource(`/ui/${utils.getGTKString()}/wispsPage.ui`);
+    dialog.getBuilder().add_from_resource(`/ui/${utils.getGTKString()}/energizeBPage.ui`);
 
     // Bind all properties.
-    dialog.bindAdjustment('wisps-animation-time');
-    dialog.bindAdjustment('wisps-scale');
-    dialog.bindColorButton('wisps-color');
+    dialog.bindAdjustment('energize-b-animation-time');
+    dialog.bindAdjustment('energize-b-scale');
+    dialog.bindColorButton('energize-b-color');
 
     // Finally, append the settings page to the main stack.
     const stack = dialog.getBuilder().get_object('main-stack');
     stack.add_titled(
-        dialog.getBuilder().get_object('wisps-prefs'), Wisps.getNick(), Wisps.getLabel());
+        dialog.getBuilder().get_object('energize-b-prefs'), EnergizeB.getNick(),
+        EnergizeB.getLabel());
   }
 
   // ---------------------------------------------------------------- API for extension.js
@@ -84,7 +84,7 @@ var Wisps = class Wisps {
   // the actor - usually windows are faded to transparency and scaled down slightly by
   // GNOME Shell. Here, we modify this behavior as well as the transition duration.
   static tweakTransitions(actor, settings) {
-    const animationTime = settings.get_int('wisps-animation-time');
+    const animationTime = settings.get_int('energize-b-animation-time');
 
     const tweakTransition = (property, value) => {
       const transition = actor.get_transition(property);
@@ -94,11 +94,10 @@ var Wisps = class Wisps {
       }
     };
 
-    // We re-target these transitions so that the window is not faded but scaled down a
-    // tiny bit.
+    // We re-target these transitions so that the window is neither scaled nor faded.
     tweakTransition('opacity', 255);
-    tweakTransition('scale-x', 0.9);
-    tweakTransition('scale-y', 0.9);
+    tweakTransition('scale-x', 1);
+    tweakTransition('scale-y', 1);
   }
 }
 
@@ -118,7 +117,7 @@ if (utils.isInShellProcess()) {
     _init(settings) {
       super._init({shader_type: Clutter.ShaderType.FRAGMENT_SHADER});
 
-      const color = Clutter.Color.from_string(settings.get_string('wisps-color'))[1];
+      const color = Clutter.Color.from_string(settings.get_string('energize-b-color'))[1];
 
       this.set_shader_source(`
 
@@ -127,89 +126,98 @@ if (utils.isInShellProcess()) {
       ${shaderSnippets.noise()}
       ${shaderSnippets.edgeMask()}
 
-      const vec2  SEED            = vec2(${Math.random()}, ${Math.random()});
-      const float WISPS_RADIUS    = 20.0;
-      const float WISPS_SPEED     = 10.0;
-      const float WISPS_SPACING   = 40 + WISPS_RADIUS;
-      const int   WISPS_LAYERS    = 8;
-      const float WISPS_IN_TIME   = 0.5;
-      const float WINDOW_OUT_TIME = 1.0;
+      const vec2  SEED         = vec2(${Math.random()}, ${Math.random()});
+      const float SHOWER_TIME  = 0.2;
+      const float SHOWER_WIDTH = 0.3;
+      const float STREAK_TIME  = 0.6;
+      const float EDGE_FADE    = 50;
+      const float SCALE        = ${settings.get_double('energize-b-scale')};
 
-      // Returns a grid of randomly moving points. Each grid cell contains one point which
-      // moves on an ellipse. 
-      float getWisps(vec2 texCoords, float gridSize, vec2 seed) {
+      // This method returns four values:
+      //  result.x: A mask for the particles which lead the shower.
+      //  result.y: A mask for the streaks which follow the shower particles.
+      //  result.z: A mask for the final "atom" particles.
+      //  result.w: The opacity of the fading window.
+      vec4 getMasks() {
+        float showerProgress = uProgress/SHOWER_TIME;
+        float streakProgress = clamp((uProgress-SHOWER_TIME)/STREAK_TIME, 0, 1);
+        float fadeProgress  = clamp((uProgress-SHOWER_TIME)/(1.0 - SHOWER_TIME), 0, 1);
 
-        // Shift coordinates by a random offset and make sure the have a 1:1 aspect ratio.
-        vec2 coords = (texCoords + hash22(seed)) * vec2(uSizeX, uSizeY);
+        // Gradient from top to bottom.
+        float t = cogl_tex_coord_in[0].t;
 
-        // Apply global scale.
-        coords /= gridSize;
+        // A smooth gradient which moves to the bottom within the showerProgress.
+        float showerMask = smoothstep(1, 0, abs(showerProgress - t - SHOWER_WIDTH) / SHOWER_WIDTH);
 
-        // Get grid cell coordinates in [0..1].
-        vec2 cellUV = mod(coords, vec2(1));
+        // This is 1 above the streak mask.
+        float streakMask = (showerProgress - t - SHOWER_WIDTH) > 0 ? 1 : 0;
 
-        // This is unique for each cell.
-        vec2 cellID = coords-cellUV + vec2(362.456);
+        // Compute mask for the "atom" particles.
+        float atomMask = getRelativeEdgeMask(0.2);
+        atomMask = max(0, atomMask - showerMask);
+        atomMask *= streakMask;
+        atomMask *= (1.0 - fadeProgress);
 
-        // Add random rotation, scale and offset to each grid cell.
-        float speed     = mix(10.0, 15.0,   hash12(cellID*seed*134.451)) / gridSize * WISPS_SPEED;
-        float rotation  = mix( 0.0,  6.283, hash12(cellID*seed*54.4129));
-        float radius    = mix( 0.5,  1.0,   hash12(cellID*seed*19.1249)) * WISPS_RADIUS;
-        float roundness = mix(-1.0,  1.0,   hash12(cellID*seed*7.51949));
+        // Make some particles visible in the streaks.
+        showerMask += 0.05 * streakMask;
 
-        vec2 offset = vec2(sin(speed * (uTime+1)) * roundness, cos(speed * (uTime+1)));
-        offset *= 0.5 - 0.5 * radius / gridSize;
-        offset = vec2(offset.x * cos(rotation) - offset.y * sin(rotation),
-                      offset.x * sin(rotation) + offset.y * cos(rotation));
+        // Add shower mask to streak mask.
+        streakMask = max(streakMask, showerMask);
 
-        cellUV += offset;
+        // Fade-out the masks at the window edges.
+        float edgeFade = getAbsoluteEdgeMask(EDGE_FADE);
+        streakMask *= edgeFade;
+        showerMask *= edgeFade;
+        
+        // Fade-out the masks from top to bottom.
+        float fade = smoothstep(0.0, 1.0, 1.0 + t - 2.0 * streakProgress);
+        streakMask *= fade;
+        showerMask *= fade;
 
-        // Use distance to center of shifted / rotated UV coordinates to draw a glaring point.
-        float dist = length(cellUV - 0.5) * gridSize / radius;
-        if (dist < 1.0) {
-          return min(10, 0.01 / pow(dist, 2.0));
-        }
+        // Compute fading window opacity.
+        float windowMask = pow(1.0 - fadeProgress, 2.0);
 
-        return 0.0;
+        return vec4(showerMask, streakMask, atomMask, windowMask);
       }
 
       void main() {
         
-        // Get the color of the window.
+        vec4 masks       = getMasks();
         vec4 windowColor = texture2D(uTexture, cogl_tex_coord_in[0].st);
-        vec4 effectColor = vec4(${color.red / 255},
-          ${color.green / 255},
-          ${color.blue / 255}, 1.0) * windowColor.a;
-        
-        // Compute several layers of moving wisps.
-        vec2 uv = (cogl_tex_coord_in[0].st-0.5) / mix(1.0, 0.5, uProgress) + 0.5;
-        uv /= ${settings.get_double('wisps-scale')};
-        float wisps = 0;
-        for (int i=0; i<WISPS_LAYERS; ++i) {
-          wisps += getWisps(uv*0.3, WISPS_SPACING, SEED * (i+1));
-        }
+        vec3 effectColor = vec3(${color.red / 255},
+                                ${color.green / 255},
+                                ${color.blue / 255});
 
-        // Compute shrinking edge mask.
-        float mask = getRelativeEdgeMask(mix(0.01, 0.5, uProgress));
+        // Dissolve window to effect color / transparency.
+        cogl_color_out = mix(vec4(effectColor, 1.0) * windowColor.a, windowColor, 0.5 * masks.w + 0.5) * masks.w;
 
-        // Compute three different progress values.
-        float wispsIn   = smoothstep(0, 1, clamp(uProgress/WISPS_IN_TIME, 0, 1));
-        float wispsOut  = smoothstep(0, 1, clamp((uProgress - WISPS_IN_TIME)/(1.0 - WISPS_IN_TIME), 0, 1));
-        float windowOut = smoothstep(0, 1, clamp(uProgress/WINDOW_OUT_TIME, 0, 1));
+        // Add leading shower particles.
+        vec2 showerUV = cogl_tex_coord_in[0].st + vec2(0, -0.7*uProgress/SHOWER_TIME);
+        showerUV *= 0.02 * vec2(uSizeX, uSizeY) / SCALE;
+        float shower = pow(simplex2D(showerUV), 10.0);
+        cogl_color_out.rgb += effectColor * shower * masks.x;
 
-        // Use a noise function to dissolve the window.
-        float noise = smoothstep(1.0, 0.0, abs(2.0 * simplex2DFractal(uv * vec2(uSizeX, uSizeY) / 250) - 1.0));
-        float windowMask = 1.0 - (windowOut < 0.5 ? mix(0.0, noise, windowOut * 2.0) : mix(noise, 1.0, windowOut * 2.0 - 1.0));
-        cogl_color_out = windowColor * windowMask * mask;
+        // Add trailing streak lines.
+        vec2 streakUV = cogl_tex_coord_in[0].st + vec2(0, -uProgress/SHOWER_TIME);
+        streakUV *= vec2(0.05 * uSizeX, 0.001 * uSizeY) / SCALE;
+        float streaks = simplex2DFractal(streakUV) * 0.5;
+        cogl_color_out.rgb += effectColor * streaks * masks.y;
 
-        // Add the wisps.
-        cogl_color_out += min(wispsIn, 1.0 - wispsOut) * wisps * effectColor * mask;
+        // Add glimmering atoms.
+        vec2 atomUV = cogl_tex_coord_in[0].st + vec2(0, -0.025*uProgress/SHOWER_TIME);
+        atomUV *= 0.2 * vec2(uSizeX, uSizeY) / SCALE;
+        float atoms = pow((simplex3D(vec3(atomUV, uTime))), 5.0);
+        cogl_color_out.rgb += effectColor * atoms * masks.z;
 
         // These are pretty useful for understanding how this works.
-        // cogl_color_out = vec4(vec3(windowMask), 1.0);
-        // cogl_color_out = vec4(vec3(wisps), 1.0);
-        // cogl_color_out = vec4(vec3(noise), 1.0);
-        // cogl_color_out = vec4(vec3(mask*min(wispsIn, 1.0 - wispsOut)), 1.0);
+        // cogl_color_out = vec4(masks.rgb, 1.0);
+        // cogl_color_out = vec4(vec3(masks.x), 1.0);
+        // cogl_color_out = vec4(vec3(masks.y), 1.0);
+        // cogl_color_out = vec4(vec3(masks.z), 1.0);
+        // cogl_color_out = vec4(vec3(masks.w), 1.0);
+        // cogl_color_out = vec4(vec3(shower), 1.0);
+        // cogl_color_out = vec4(vec3(streaks), 1.0);
+        // cogl_color_out = vec4(vec3(atoms), 1.0);
       }
       `);
     };